grpc-server/requirements.txt
googleapis-common-protos==1.5.5
grpcio==1.16.0
grpcio-tools==1.16.0
protobuf==3.6.1
six==1.11.0
